Refactor: Chat 컴포넌트 prop 수정
